backend/requirements.txt
PyMuPDF>=1.24.0
opencv-python>=4.9.0
google-generativeai>=0.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
